A few more DL tweaks
